refactor: general fixes
