+++ apps/frontend/src/app/(landing)/@experience/components/experience-card/employement-card.tsx
@@ -1,4 +1,4 @@
-import { cn } from '../../../../../lib/utils';
+import { cn } from '@/lib/utils/cn';
 import Image from 'next/image';
 import Link from 'next/link';
 import { IoLocationOutline } from 'react-icons/io5';
