+++ apps/frontend/src/app/(landing)/@experience/components/experience-list.tsx
@@ -1,4 +1,4 @@
-import { cn } from '../../../../lib/utils';
+import { cn } from '@/lib/utils/cn';
 import { ReactNode } from 'react';
 
 type Props = {
