+++ apps/frontend/src/lib/utils/name-letters.ts
@@ -1,0 +1,7 @@
+export function nameLetters(name: string) {
+  return name
+    .split(' ')
+    .map((word) => word[0])
+    .join('')
+    .toUpperCase();
+}
